Added commentary section
