+++ src/app/screens/Detail/components/DataSection/constants.ts
