+++ src/constants/colors.ts
@@ -2,4 +2,6 @@ export const LIGHT_BLUE = '#EAF6FA';
 export const LIGHT_BLUE_1 = '#00ADEE';
 export const WHITE = '#FFFFFF';
 export const GREY = '#4A4A4A';
+export const DARK_GREY = '#030303';
+export const LIGHT_GREY = '#C8C7CC';
 export const GREEN = '#A5CD39';
