+++ src/constants/commentaries.ts
@@ -1,0 +1,39 @@
+import { CommentProps } from '@interfaces/commentaries';
+import avatar1 from '@assets/img_user1.png';
+import avatar2 from '@assets/img_user2.png';
+
+export const COMMENTS: CommentProps[] = [
+  {
+    id: 1,
+    avatar: avatar1,
+    author: 'John Appleseed',
+    content:
+      "It was great to see you again earlier. Let's definitely get that coffe It was great to see you again earlier."
+  },
+  {
+    id: 2,
+    author: 'Susan Collins',
+    avatar: avatar2,
+    content: "Let's definitely get that coffe It was great to see you again earlier."
+  },
+  {
+    id: 3,
+    avatar: avatar1,
+    author: 'John Appleseed',
+    content:
+      "It was great to see you again earlier. Let's definitely get that coffe It was great to see you again earlier."
+  },
+  {
+    id: 4,
+    author: 'Susan Collins',
+    avatar: avatar2,
+    content: "Let's definitely get that coffe It was great to see you again earlier."
+  },
+  {
+    id: 5,
+    avatar: avatar1,
+    author: 'John Appleseed',
+    content:
+      "It was great to see you again earlier. Let's definitely get that coffe It was great to see you again earlier."
+  }
+];
